check a baseline and mixture file together to ensure the known_collections are valid if they exist

Simple function that checks for known_collections columns in a reference and mixture and makes sure that they are compliant. If there is a non-NA entry in the Mixture frame's known_collection column this function returns TRUE. Otherwise it returns FALSE.

Usage

check_known_collections(R, M)

Arguments

R

reference data frame

M

mixture data frame