Calculate a matrix of sum-of-squares of genotype log-likelihoods for a genetic dataset

Takes a list of key parameters from a genetic dataset, and calculates the sum of squared log-likelihood of each individual's genotype, given the allele counts in each collection. This is used for the quick-and-dirty Z-score calculations.

Usage

geno_logL_ssq(par_list)

Arguments

par_list

genetic data converted to the param_list format by tcf2param_list

Value

geno_logL returns a matrix with C rows and I columns. Each column represents an individual, and each row the log-likelihood of that individual's genotype, given the allele counts in that collection

Details

Leave-One-Out cross-validation is used to avoid bias in log-likelihood for an individual's known collection of origin

Examples

example(tcf2param_list)
#> 
#> tcf2p_> # after adding support for haploid markers we need to pass
#> tcf2p_> # in the ploidies vector.  These markers are all diploid...
#> tcf2p_> locnames <- names(alewife)[-(1:16)][c(TRUE, FALSE)]
#> 
#> tcf2p_> ploidies <- rep(2, length(locnames))
#> 
#> tcf2p_> names(ploidies) <- locnames
#> 
#> tcf2p_> ale_par_list <- tcf2param_list(alewife, 17, ploidies = ploidies)
#> Summary Statistics:
#> 
#> 1070 Individuals in Sample
#> 
#> 11 Loci: Aa046, Aa070, Aa074, Aa081, Aa091, Aa093, Ap010, Ap033, Ap038, Ap058, Ap071
#> 
#> 3 Reporting Units: NNE, SNE, MAT
#> 
#> 21 Collections: EMA, STG, PIS, MYS, MON, TBR, GIL, THA, BRI, CON, QUI, HOU, PEQ, MIA, HUD, DEL, NAN, RAP, CHO, ROA, ALL
#> 
#> 6.31% of allelic data identified as missing
ale_glL <- geno_logL(ale_par_list)