From the pattern of missing data at each individual, compute the expected mean and variance of the logl

This takes a param_list so that it has access to individual genotypes (and hence can cycle through them and know which are missing and which are not.) It also takes a matrix of per-locus logl means and variances like what is computed by per_locus_means_and_vars.

Usage

rcpp_indiv_specific_logl_means_and_vars(par_list, MV)

Arguments

par_list

genetic data converted to the param_list format by tcf2param_list

MV

a list of two matrices, one of means and the other of variances, which are C x L matrices. This is basically the list that is returned by per_locus_means_and_vars.

Value

a matrix with C rows and I columns. Each row represents a collection, and each column an individual.

Details

This function doesn't do any checking to assure that the par_list and the per-locus logl means matrix are made for one another. (i.e. use the same collections in the same order.)