rubias: Bayesian inference from the conditional genetic stock identification model

Read the "rubias-overview" vignette for information on data input formats and how to use the package

the rubias main high-level functions

The following functions are wrappers, designed for user-friendly input and useful output:

infer_mixture is used to perform genetic stock identification. Options include standard MCMC and the parametric bootstrap bias correction.

self_assign does simple self-assignment of individuals in a reference data set to the various collections in the reference data set.

assess_reference_loo does leave-one-out based simulations to predict how accurately GSI can be done.

assess_reference_mc uses Monte-Carlo cross-validation based simulations to predict how accurately GSI can be done.

assess_pb_bias_correction attempts to demonstrate how much (or little) improvement can be expected from the parametric bootstrap correction given a particular reference data set.

genetic data format

See the vignette.

example data

alewife, blueback, and chinook are genetic data sets that are useful for playing around with rubias and testing it out.

See also

Author

Maintainer: Eric C. Anderson eriq@rams.colostate.edu

Authors:

Ben Moran moran.ben@husky.neu.edu